requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==3.2.2
cffi==2.0.0
click==8.3.1
ecdsa==0.19.1
fastapi==0.128.0
filelock==3.20.3
fsspec==2026.1.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.1
opencv-python==4.13.0.90
pandas==3.0.0
passlib==1.7.4
pillow==12.1.0
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-jose==3.5.0
python-multipart==0.0.22
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.0
setuptools==80.10.2
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.50.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
